+++ TrainingPlanner/app/src/main/java/com/example/trainingplanner/ui/repositories/WorkoutsRepository.kt
@@ -29,7 +29,7 @@ object WorkoutsRepository {
     suspend fun createWorkout(
         trainingPlanCode: String,
         date: LocalDate,
-        amount: Int,
+        amount: Float,
         unit: String,
         type: String,
         description: String,
